Update to Jest 20
