+++ package.json
@@ -13,7 +13,7 @@
     "eslint-plugin-promise": "^3.5.0",
     "eslint-plugin-standard": "^3.0.1",
     "fetch-mock": "^5.10.1",
-    "jest": "^18.0.0",
+    "jest": "^20.0.0",
     "jest-fetch-mock": "^1.1.1",
     "lint-staged": "^3.4.2",
     "mock-local-storage": "^1.0.2",
