+++ src/__tests__/LoginAvatar-test.js
@@ -1,0 +1,39 @@
+import React from "react";
+import { shallow } from "enzyme";
+import { LoginAvatar } from "../LoginAvatar";
+import {
+  mountWithIntl,
+  shallowWithIntl
+} from "../helpers/intl-enzyme-test-helper.js";
+import intlProps from "../helpers/intl-props-test-helper.js";
+
+function setup() {
+  const props = {
+    isAuthenticated: true,
+    dispatch: jest.fn(),
+    intl: intlProps,
+    profile: {}
+  };
+
+  const enzymeWrapper = shallowWithIntl(<LoginAvatar {...props} />);
+
+  return {
+    props,
+    enzymeWrapper
+  };
+}
+
+describe("components", () => {
+  describe("LoginAvatar", () => {
+    it("should render user menu if logged in", () => {
+      const { enzymeWrapper } = setup();
+      expect(enzymeWrapper.find("div").hasClass("avatar")).toBe(true);
+    });
+
+    it("should show login button if logged out", () => {
+      const { enzymeWrapper } = setup();
+      enzymeWrapper.setProps({ isAuthenticated: false });
+      expect(enzymeWrapper.find("div").hasClass("loginButton")).toBe(true);
+    });
+  });
+});
